requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-community
langchain-ollama
langchain-text-splitters

faiss-cpu
pypdf
